+++ sw/ai_accelerator_eigen_integers/src/nn/relu.cpp
@@ -1,15 +1,15 @@
 #include <Eigen/Dense>
 #include "relu.hpp"
 
-Eigen::MatrixXf nn::ReLU::forward(const Eigen::MatrixXf& input) {
+Eigen::MatrixXi nn::ReLU::forward(const Eigen::MatrixXi& input) {
     this->input = input;
     // Set all negative values to zero
-    Eigen::MatrixXf output = input.cwiseMax(0.0f);
+    Eigen::MatrixXi output = input.cwiseMax(0);
     return output;
 }
 
-Eigen::MatrixXf nn::ReLU::backward(const Eigen::MatrixXf& gradOutput) {
+Eigen::MatrixXi nn::ReLU::backward(const Eigen::MatrixXi& gradOutput) {
     // Gradient is 1 for positive input values, 0 otherwise
-    Eigen::MatrixXf gradInput = (input.array() > 0).select(gradOutput, 0);
+    Eigen::MatrixXi gradInput = (input.array() > 0).select(gradOutput, 0);
     return gradInput;
 }
