+++ sw/ai_accelerator_eigen_integers/src/nn/relu.hpp
@@ -6,14 +6,14 @@
 namespace nn{
     class ReLU : public Module {
     public:
-        Eigen::MatrixXf forward(const Eigen::MatrixXf& input) override;
-        Eigen::MatrixXf backward(const Eigen::MatrixXf& gradOutput) override;
+        Eigen::MatrixXi forward(const Eigen::MatrixXi& input) override;
+        Eigen::MatrixXi backward(const Eigen::MatrixXi& gradOutput) override;
         void update(double learningRate) override {} // No update needed for ReLU as it is not learnable
         std::string toString() override { return "ReLU"; }
         std::string getName() override { return "ReLU"; }
         virtual void serialize(FILE* file) const override {};
         virtual void deserialize(FILE* file) override {};
     private:
-        Eigen::MatrixXf input;
+        Eigen::MatrixXi input;
     };
 }
